+++ src/main/java/com/navi/dcim/task/TaskRepository.java
@@ -3,11 +3,10 @@ package com.navi.dcim.task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 
     List<Task> findByActive(boolean active);
 }
